ECCOtour

ECCOtour.Nino34file

ECCOtour.allstats

ECCOtour.annualcontrib!

ECCOtour.apply_regional_mask!

ECCOtour.centerlon!

ECCOtour.columnscale!

ECCOtour.columnscale!

ECCOtour.depthlevels

ECCOtour.extract_sst34

ECCOtour.factors4regularpoles

ECCOtour.factors4velocity

ECCOtour.faststats

ECCOtour.get_filtermatrix

ECCOtour.hanncoeffs

ECCOtour.hannfilter

ECCOtour.hannfilter

ECCOtour.hannsum

ECCOtour.hannsum!

ECCOtour.hannsum!

ECCOtour.historicalNino34

ECCOtour.inrectangle

ECCOtour.latlonC

ECCOtour.latlonG

ECCOtour.matrixfilter

ECCOtour.matrixspray

ECCOtour.mdsio2sigma1

ECCOtour.mean

ECCOtour.mean

ECCOtour.mean

ECCOtour.ncwritefromtemplate

ECCOtour.netcdf2dict

ECCOtour.netcdf2dict

ECCOtour.netcdf2sigma1

ECCOtour.patchmean

ECCOtour.position_label

ECCOtour.prereginterp

ECCOtour.pressurelevels

ECCOtour.readNino34

ECCOtour.readarea

ECCOtour.reginterp

ECCOtour.regional_mask

ECCOtour.remove_climatology

ECCOtour.remove_seasonal

ECCOtour.remove_seasonal

ECCOtour.replace!

ECCOtour.replace!

ECCOtour.replace!

ECCOtour.replace!

ECCOtour.replace!

ECCOtour.searchdir

ECCOtour.seasonal_matrices

ECCOtour.setupLLCgrid

ECCOtour.sigma

ECCOtour.sigma1column

ECCOtour.sigma1grid

ECCOtour.sigmacolumn

ECCOtour.std

ECCOtour.time_label

ECCOtour.timestamp_monthly_v4r4

ECCOtour.trend_matrices

ECCOtour.trend_theta

ECCOtour.trend_theta!

ECCOtour.var2regularpoles

ECCOtour.var2sigmacolumn

ECCOtour.vars2regularpoles

ECCOtour.vars2sigma1

ECCOtour.vars2sigma1

ECCOtour.wrapdist

ECCOtour.write_vars

ECCOtour.writeregularpoles

ECCOtour.writeregularpoles

ECCOtour.writeregularpoles

ECCOtour.writeregularpoles

ECCOtour.Nino34file — Method

function Nino34file()
Get location of the historical Nino34 Google Drive file
Download if necessary
FUTURE UPDATES: use GoogleDrive package

Output

fileloc: local location of gdrive file

ECCOtour.allstats — Method

function allstats(x)
Compute fast statistics of gcmgrid type using function calls, same as faststats

Input

x::MeshArrays.gcmarray{Float32,1,Array{Float32,2}}: input of gcmarray type

Output

xbar::Float32: mean value after discarding dryval

xmax::Float32: maximum value

xmin::Float32: minimum value

σx::Float32: standard deviation

absxbar::Float32: mean of absolute value

ECCOtour.annualcontrib! — Method

function matrixfilter_annualcontrib!(θout,F,θin)

Arguments

θout: filtered values

F: filter function as a matrix

θin: input values at all times t (e.g., θ(t))

ECCOtour.apply_regional_mask! — Method

function regional_mask(field,mask)

Arguments

field: input field that is mutated/modified

mask: multiplicative factor, same spatial grid as field

ECCOtour.centerlon! — Method

function centerlon!
Make a longitudinal coordinate system
centered at lonmid.
Makes handling wraparound easy later on.

ECCOtour.columnscale! — Method

function columnscale!(product,M,flux)

product::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}

M::Array{Float32,1}

flux::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}

ECCOtour.columnscale! — Method

function columnscale!(product,M,flux)

product::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}

M::Array{Float32,1}

flux::MeshArrays.gcmarray{Float32,1,Array{Float32,2}}

ECCOtour.depthlevels — Method

function depthlevels(γ)
Depths of ECCO v4r4 grid (positive vals)

ECCOtour.extract_sst34 — Method

function extract_sst34
extract by reading multiple files

ECCOtour.factors4regularpoles — Method

function factors4regularpoles(γ)
Get interpolation factors for regularpoles grid in one place

ECCOtour.factors4velocity — Method

function factors4regularpoles(γ)
Get interpolation factors for regularpoles grid in one place

ECCOtour.faststats — Method

function faststats(x)
Compute fast statistics of gcmgrid type using function calls, eliminate redundancy

Input

x::MeshArrays.gcmarray{Float32,1,Array{Float32,2}}: input of gcmarray type

Output

xbar::Float32: mean value after discarding dryval

xmax::Float32: maximum value

xmin::Float32: minimum value

σx::Float32: standard deviation

absxbar::Float32: mean of absolute value

ECCOtour.get_filtermatrix — Method

get_filtermatrix(tin,tout)

Arguments

tin: input times

tout: output times

Output

Eout2in: matrix that maps from tout to tin

Fin2out: matrix that maps from tin to tout

ECCOtour.hanncoeffs — Method

hanncoeffs(tin,tout,L)

Arguments

tin: input dataset time (vector valued)

tout: time where filtered output will be created

L: filter length

Output

h: normalized coefficients of Hann(ing) filter

ECCOtour.hannfilter — Method

function hannfilter(θin,tin,tout,L,γ)

Arguments

θin::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}

tin::StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}

tout::StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}

L::Float64

γ::gcmgrid

Output

θout

ECCOtour.hannfilter — Method

function hannfilter(θin,tin,tout,L)

Arguments

θin::Array{Float32,1}: input

tin

tout

L

Output

θout: output

ECCOtour.hannsum! — Method

hannsum!(θout,θin,tin,tout,L)
use this when a timeseries is available

Arguments

θout: output value at time th (e.g., θh(th))

θin: input value at all times t (e.g., θ(t))

tin: input dataset time (vector valued)

tout: time where filtered output will be created

L: filter length

ECCOtour.hannsum! — Method

function hannsum!(θout,θin,tin,tout,toutindex,L::Float64)

 Call this version for mesharrays

Arguments

θout::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}: output value at time tout (e.g., θout(tout))

θin::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}: input value at all times t (e.g., θ(t))

tin::StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}: input dataset time (vector valued)

tout::StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}: time where filtered output will be created

toutindex::Integer: output times?

L::Float64: filter length

ECCOtour.hannsum — Method

hannsum(θin,tin,tout,L)
use this when a timeseries is available

Arguments

θ: input value at all times t (e.g., θ(t))

tout: time where filtered output will be created

tin: input dataset time (vector valued)

L: filter length

Output

θout: output value at time th (e.g., θh(th))

ECCOtour.historicalNino34 — Method

function historicalNino34(baselineyears)
Get historical Nino3.4 data and SST climatology from HadISST 1

Argument

baselineyears: for computation of SST climatology, i.e., (1960,2020)

Output

nino34: historical Nino3.4 index

tnino34: time in years CE corresponding to index

SSTclimatology: monthly mean values in Nino3.4 patch

ECCOtour.inrectangle — Method

function inrectangle(lonin,latin,lons,lats)
find all gridcells within Cartesian rectangle

Arguments

lons: tuple of longitude range

lats: tuple of latitude range

Output

rectangle: boolean of type gcmarray

ECCOtour.latlonC — Method

function latlonC(γ)
Latitude-longitude of ECCOv4r4 "C" (tracer) grid

ECCOtour.latlonG — Method

function latlongG(γ)
Latitude-longitude of ECCOv4r4 "G" (velocity) grid

ECCOtour.matrixfilter — Method

matrixfilter(F,froot,years,γ)
writing it in a funny way to save computation
issue with timeseries being read in different files

Arguments

F: filter in matrix form

froot: filename root

years: iterator for multiple files

γ: GCM grid (meshArray type)

ECCOtour.matrixspray — Method

matrixspray(F,rmfield,frootin,frootout,years,γ)

writing it in a funny way to save computation
issue with timeseries being read in different files

Arguments

F: spray operator in matrix form

rmfield: field to be removed from files

frootin: filename root of input field

frootout: filename root of output field

years: iterator for multiple files

γ: GCM grid (meshArray type)

ECCOtour.mdsio2sigma1 — Method

function mdsio2sigma1
Take variables in a filelist, read, map to sigma1, write to file.

ECCOtour.mean — Method

function mean(x::MeshArrays.gcmarray{Float32,1,Array{Float32,2}},dryval::Float32)
Compute mean of gcmgrid type using function calls, eliminate redundancy
Eliminate all values = dryval
Avoid a name clash with StatsBase and MeshArrays

ECCOtour.mean — Method

function mean(x::MeshArrays.gcmarray{Float32,1,Array{Float32,2}},dryval::Float32)
Compute area-weighted mean of gcmgrid type using filtered with function isgood
Area weighting = area
Eliminate all values = isgood(x) -> true
Avoid a name clash with StatsBase and MeshArrays

ECCOtour.mean — Method

function mean(x::MeshArrays.gcmarray{Float32,1,Array{Float32,2}},dryval::Float32)
Compute area-weighted mean of gcmgrid type using function calls, eliminate redundancy
Area weighting = area
Eliminate all values = dryval
Avoid a name clash with StatsBase and MeshArrays

ECCOtour.ncwritefromtemplate — Method

function ncwritefromtemplate This version writes NetCDF output on the regularpoles grid

ECCOtour.netcdf2dict — Method

netcdf2dict array output (?)

ECCOtour.netcdf2dict — Method

netcdf2dict gcmarray output

ECCOtour.netcdf2sigma1 — Method

function netcdf2sigma1
Take variables in a NetCDF filelist, read, map to sigma1, write to file.

ECCOtour.patchmean — Method

 function patchmean(x,area,ϕ,λ,ispatch,iswet)
 get weight for rectangle region.

Arguments

x: variable of interest

area: weighting

ϕ: lat

λ: lon

ispatch: true in patch of interest

iswet = function that is true if in ocean

Output

xbar: weighted filtered average of x

ECCOtour.position_label — Method

function position_label(lon,lat)

Arguments

lon: longitude

lat: latitude

Output

lbl: LaTex string used for figure labels

ECCOtour.prereginterp — Method

function prereginterp(latgrid,longrid,γ)
prepare for regular interpolation
regular = onto rectangular 2d map

Arguments

latgrid: 1d array of latitude

longrid: 1d array of longitude

γ: GCM grid

Output

f,i,j,w: interpolation factors

ECCOtour.pressurelevels — Method

function pressurelevels(z)
Standard pressures of ECCO v4r4 grid

ECCOtour.readNino34 — Method

function readNino34()
Get historical Nino3.4 data from HadISST 1

Output

SST_nino34: local location of gdrive file

ECCOtour.readarea — Method

function readarea(γ)
area of ECCO v4r4 grid

ECCOtour.reginterp — Method

function reginterp(fldin,nx,ny,f,i,j,w)
regular interpolation with precomputed factors
regular = onto rectangular 2d map

Arguments

fldin: gcmarray field of input

nx,ny: x,y dimension of output regular field

f,i,j,w: precomputed interpolation factors

Output

fldout: interpolated regular 2d field

ECCOtour.regional_mask — Method

function regional_mask(latpt,lonpt,latrect,lonrect,dlat,dlon)

Arguments

latpt: latitude grid

lonpt: longitude grid

latrect: tuple of latitude range of central rectangle

lonrect: tuple of longitude range of central rectangle

dlon: width in degrees longitude of East/West fade zones

'dlat': width in degrees latitude of North/South fade zones

Output

'mask': space and time field of surface forcing, value of zero inside

designated lat/lon rectangle and fading to 1 outside sponge zone on each edge. This is because this field ends up being SUBTRACTED from the total forcing

ECCOtour.remove_climatology — Method

function remove_climatology(x,xbar)
`x` = long monthly timeseries, starting in Jan
`xbar` = 12-month climatology, starting in Jan
remove climatology

ECCOtour.remove_seasonal — Method

function remove_seasonal(x,Ecycle,Fcycle,γ) natively
`x` = gcmarray of long monthly timeseries, starting in Jan
remove seasonal cycle of this timeseries

ECCOtour.remove_seasonal — Method

function remove_seasonal(x,Ecycle,Fcycle,γ) natively
`x` = long monthly timeseries, starting in Jan
remove seasonal cycle of this timeseries

ECCOtour.replace! — Method

function replace!(f,a::MeshArrays.gcmarray{Float64,1,Array{Float64,2}}}

ECCOtour.replace! — Method

function replace!(a::MeshArrays.gcmarray{Float32,1,Array{Float32,2}}},b::Pair)

ECCOtour.replace! — Method

function replace!(a::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}},b::Pair)

ECCOtour.replace! — Method

function replace!(a::MeshArrays.gcmarray{Float64,1,Array{Float64,2}}},b::Pair)

ECCOtour.replace! — Method

function replace!(a::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}},b::Pair)

ECCOtour.searchdir — Method

function searchdir(path,key)
a useful one-line function

Arguments

path: directory to search for file

key: expression to search for in path

ECCOtour.seasonal_matrices — Function

function seasonal_matrices(fcycle,t,overtones=1)

Arguments

fcycle: frequency of seasonal cycle

t: time

overtones=1: optional argument for number of overtones

Output

E: matrix that solves E*parameters= seasonal cycle

F=E†: generalized inverse of E

ECCOtour.setupLLCgrid — Method

function setupLLCgrid(path_grid)
Load the ECCOv4r4 grid at location path_grid into a MeshArray variable

Arguments

path_grid: directory to search for grid files

Output

γ::gcmgrid: MITgcm grid

ECCOtour.sigma — Method

function sigma(θ,S,pz,p₀)
sigma values from SeaWaterDensity for gcmarrays

Arguments

θ::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}: potential temperature

S::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}: practical salinity

pz::Array{Float64,1} : vertical profile of standard pressures

p₀::Int64 : reference pressure for sigma value

Output

σ::MeshArrays.gcmarray{Float32,2,Array{Float32,2}}: potential density referenced to p₀ minus 1000 kg/m³

ECCOtour.sigma1column — Method

function sigma1column(θ,S,p)
σ₁ for a water column

Arguments

θz::Array{Float,1}}: potential temperature

Sz::Array{Float,1}}: practical salinity

pz::Array{Float,1}: vertical profile of standard pressures

Output

σ₁: sigma-1 for wet points in column

ECCOtour.sigma1grid — Method

function sigma1grid() Standard choice of sigma1 surfaces

ECCOtour.sigmacolumn — Method

function sigmacolumn(θ,S,p,p0)
σ for a water column

Arguments

θz::Array{Float,1}}: potential temperature

Sz::Array{Float,1}}: practical salinity

pz::Array{Float,1}: vertical profile of standard pressures

p0::Float: reference pressure

Output

σ: sigma for wet points in column

ECCOtour.std — Method

function std(x::MeshArrays.gcmarray{Float32,1,Array{Float32,2}},xbar::Float32,dryval)
Compute standard deviation of gcmgrid type using function calls, eliminate redundancy
Eliminate all values = dryval
Avoid a name clash with StatsBase and MeshArrays

ECCOtour.time_label — Method

function time_label(index)
Time label useful in plots. For ECCOv4r4.

Input

index::Integer: index number for monthly average field, number of months since Jan. 1992

Output

tlbl: label with month name (abbrev. with 3 letters) and calendar year CE

ECCOtour.timestamp_monthly_v4r4 — Method

function timestamp_monthly_v4r4(t)
print year and month given time index
assuming using ECCOv4r4

ECCOtour.trend_matrices — Method

function trend_matrices(t)

Arguments

t: time

Output

E: matrix that solves E*parameters= timeseries

F=E†: generalized inverse of E

ECCOtour.trend_theta! — Method

function trend_theta!(β,diagpathexp,path_out,tecco,γ,F)
get linear trend of potential temperature
from monthly-average ECCOv4r4 gcmarray fields

Arguments

β: gcmarray field of trends

diagpathexp: where to find files

tecco: time stamps of monthly output

γ: GCM grid information

F: linear estimator of trend

Output

β: updated trends

ECCOtour.trend_theta — Method

function trend_theta(diagpathexp,path_out,tecco,γ,F)
get linear trend of potential temperature
from monthly-average ECCOv4r4 gcmarray fields

Arguments

diagpathexp: where to find files

tecco: time stamps of monthly output

γ: GCM grid information

F: linear estimator of trend

Output

β: updated trends

ECCOtour.var2regularpoles — Method

var2regularpoles Take one gcmarray in memory, put on regularpoles grid

ECCOtour.var2sigmacolumn — Method

function var2sigmacolumn(σ,v,sig1grid,splorder)
map θ,S, p onto σ₁ surfaces for a water column

Arguments

σ::Array{Float,1}}`: sigma values of input variable

v::Array{Float,1}}: variable of interest

sig1: σ surface values

splorder: 1-5, order of spline

Output

θonσ: variable on sig1 sigma surfaces

ECCOtour.vars2regularpoles — Method

function vars2regularpoles(vars::Dict{String,MeshArrays.gcmarray{Float32,1,Array{Float32,2}}},γ,nx,ny,nyarc,farc,iarc,jarc,warc,nyantarc,fantarc,iantarc,jantarc,wantarc) variables interpolated onto regularpoles grid

ECCOtour.vars2sigma1 — Method

function vars2sigma1(vars,p,sig1grid,γ,spline_order)
map variables from regularpoles grid onto sigma1 surfaces

Arguments

vars::Dict{String,Array{Float64,3}}}: dict of 3d arrays

p::Array{Float64,1} : vertical profile of standard pressures

sig1grid: σ₁ surface values

γ: grid description needed for preallocation

splorder: 1-5, order of spline

Output

varsσ::Dict{String,MeshArrays.gcmarray{Float32,2,Array{Float32,2}}}: dict of gcmarrays of variables on sigma1 surfaces

ECCOtour.vars2sigma1 — Method

function vars2sigma1(vars,p,sig1grid,γ,spline_order)
map variables onto sigma1 surfaces for gcmarrays

Arguments

vars::Dict{String,MeshArrays.gcmarray{Float32,2,Array{Float32,2}}}: dict of gcmarrays

p::Array{Float64,1} : vertical profile of standard pressures

sig1grid: σ₁ surface values

γ: grid description needed for preallocation

splorder: 1-5, order of spline

Output

varsσ::Dict{String,MeshArrays.gcmarray{Float32,2,Array{Float32,2}}}: dict of gcmarrays of variables on sigma1 surfaces

ECCOtour.wrapdist — Method

function wrapdist
Longitudinal distance in degrees
Passing the date line may be the shortest distance.

ECCOtour.write_vars — Method

function write_vars This version writes mdsio output on the native grid

ECCOtour.writeregularpoles — Method

function writeregularpoles(vars,γ,pathout,filesuffix,filelog,λC,lonatts,ϕC,latatts,z,depthatts)

ECCOtour.writeregularpoles — Method

function writeregularpoles(vars,γ,pathout,filesuffix,filelog,λC,lonatts,ϕC,latatts,z,depthatts)

ECCOtour.writeregularpoles — Method

function writeregularpoles(vars,pathout,filesuffix,filelog,λC,lonatts,ϕC,latatts)

ECCOtour.writeregularpoles — Method

function writeregularpoles(vars,pathout,filesuffix,filelog,λC,lonatts,ϕC,latatts)